Move edr tests to own file
