+++ tests/test_api.py
@@ -46,7 +46,6 @@ from pygeoapi.api import (
     validate_subset, F_HTML, F_JSON, F_JSONLD, F_GZIP, __version__, apply_gzip,
 )
 from pygeoapi.api.coverages import get_collection_coverage
-from pygeoapi.api.environmental_data_retrieval import get_collection_edr_query
 from pygeoapi.api.itemtypes import (
     get_collection_schema, get_collection_queryables, get_collection_item,
     get_collection_items, manage_collection_item)
@@ -1591,193 +1590,6 @@ def test_get_collection_tiles(config, api_):
     content = json.loads(response)
     assert len(content['links']) > 0
     assert len(content['tilesets']) > 0
-
-
-def test_get_collection_edr_query(config, api_):
-    # edr resource
-    req = mock_api_request()
-    rsp_headers, code, response = api_.describe_collections(req, 'icoads-sst')
-    collection = json.loads(response)
-    parameter_names = list(collection['parameter_names'].keys())
-    parameter_names.sort()
-    assert len(parameter_names) == 4
-    assert parameter_names == ['AIRT', 'SST', 'UWND', 'VWND']
-
-    # no coords parameter
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # bad query type
-    req = mock_api_request({'coords': 'POINT(11 11)'})
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'corridor')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # bad coords parameter
-    req = mock_api_request({'coords': 'gah'})
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # bad parameter_names parameter
-    req = mock_api_request({
-        'coords': 'POINT(11 11)', 'parameter_names': 'bad'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # all parameters
-    req = mock_api_request({'coords': 'POINT(11 11)'})
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    data = json.loads(response)
-
-    axes = list(data['domain']['axes'].keys())
-    axes.sort()
-    assert len(axes) == 3
-    assert axes == ['TIME', 'x', 'y']
-
-    assert data['domain']['axes']['x']['start'] == 11.0
-    assert data['domain']['axes']['x']['stop'] == 11.0
-    assert data['domain']['axes']['y']['start'] == 11.0
-    assert data['domain']['axes']['y']['stop'] == 11.0
-
-    parameters = list(data['parameters'].keys())
-    parameters.sort()
-    assert len(parameters) == 4
-    assert parameters == ['AIRT', 'SST', 'UWND', 'VWND']
-
-    # single parameter
-    req = mock_api_request({
-        'coords': 'POINT(11 11)', 'parameter_names': 'SST'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    data = json.loads(response)
-
-    assert len(data['parameters'].keys()) == 1
-    assert list(data['parameters'].keys())[0] == 'SST'
-
-    # Zulu time zone
-    req = mock_api_request({
-        'coords': 'POINT(11 11)',
-        'datetime': '2000-01-17T00:00:00Z/2000-06-16T23:00:00Z'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    # bounded date range
-    req = mock_api_request({
-        'coords': 'POINT(11 11)',
-        'datetime': '2000-01-17/2000-06-16'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    data = json.loads(response)
-    time_dict = data['domain']['axes']['TIME']
-
-    assert time_dict['start'] == '2000-02-15T16:29:05.999999999'
-    assert time_dict['stop'] == '2000-06-16T10:25:30.000000000'
-    assert time_dict['num'] == 5
-
-    # unbounded date range - start
-    req = mock_api_request({
-        'coords': 'POINT(11 11)',
-        'datetime': '../2000-06-16'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    data = json.loads(response)
-    time_dict = data['domain']['axes']['TIME']
-
-    assert time_dict['start'] == '2000-01-16T06:00:00.000000000'
-    assert time_dict['stop'] == '2000-06-16T10:25:30.000000000'
-    assert time_dict['num'] == 6
-
-    # unbounded date range - end
-    req = mock_api_request({
-        'coords': 'POINT(11 11)',
-        'datetime': '2000-06-16/..'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    data = json.loads(response)
-    time_dict = data['domain']['axes']['TIME']
-
-    assert time_dict['start'] == '2000-06-16T10:25:30.000000000'
-    assert time_dict['stop'] == '2000-12-16T01:20:05.999999996'
-    assert time_dict['num'] == 7
-
-    # some data
-    req = mock_api_request({
-        'coords': 'POINT(11 11)', 'datetime': '2000-01-16'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.OK
-
-    # no data
-    req = mock_api_request({
-        'coords': 'POINT(11 11)', 'datetime': '2000-01-17'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.NO_CONTENT
-
-    # position no coords
-    req = mock_api_request({
-        'datetime': '2000-01-17'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'position')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # cube bbox parameter 4 dimensional
-    req = mock_api_request({
-        'bbox': '0,0,10,10'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'cube')
-    assert code == HTTPStatus.OK
-
-    # cube bad bbox parameter
-    req = mock_api_request({
-        'bbox': '0,0,10'
-    })
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'cube')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # cube no bbox parameter
-    req = mock_api_request({})
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'icoads-sst', None, 'cube')
-    assert code == HTTPStatus.BAD_REQUEST
-
-    # cube decreasing latitude coords and S3
-    req = mock_api_request({
-        'bbox': '-100,40,-99,45',
-        'parameter_names': 'tmn',
-        'datetime': '1994-01-01/1994-12-31',
-    })
-
-    rsp_headers, code, response = get_collection_edr_query(
-        api_, req, 'usgs-prism', None, 'cube')
-    assert code == HTTPStatus.OK
-
 
 def test_validate_bbox():
     assert validate_bbox('1,2,3,4') == [1, 2, 3, 4]
